+++ lib/json_api_client/middleware/factory.ex
@@ -1,10 +1,11 @@
 defmodule JsonApiClient.Middleware.Factory do
-  @moduledoc """
-  Provides all configured Middlewares.
-  """
+  @moduledoc false
 
   def middlewares do
-    configured_middlewares() ++ [{JsonApiClient.Middleware.HTTPClient, nil}]
+    configured_middlewares() ++ [
+      {JsonApiClient.Middleware.DocumentParser, nil},
+      {JsonApiClient.Middleware.HTTPClient, nil}
+    ]
   end
 
   defp configured_middlewares do