+++ test/json_api_client/middleware/fuse_test.exs
@@ -3,7 +3,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
   doctest JsonApiClient.Middleware.Fuse, import: true
 
   import Mock
-  alias JsonApiClient.Request
+  alias JsonApiClient.{Request, Response, RequestError}
   alias JsonApiClient.Middleware.Fuse
 
   @request %Request{}
@@ -24,7 +24,10 @@ defmodule JsonApiClient.Middleware.FuseTest do
         }
       ]
       ) do
-        assert {:error, %{reason: "Unavailable - json_api_client circuit blown"}} =
+        assert {:error, %RequestError{
+          message: "Unavailable - json_api_client circuit blown",
+          original_error: "Unavailable - json_api_client circuit blown",
+          status: nil}} =
         Fuse.call(@request, fn request ->
           Agent.update(agent, fn count -> count + 1 end)
           assert request == @request
@@ -49,6 +52,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
       Fuse.call(@request, fn request ->
         Agent.update(agent, fn count -> count + 1 end)
         assert request == @request
+        {:ok, %Response{}}
       end, [])
 
       assert Agent.get(agent, fn count -> count end) == 1
@@ -69,6 +73,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
       Fuse.call(@request, fn request ->
         Agent.update(agent, fn count -> count + 1 end)
         assert request == @request
+        {:ok, %Response{}}
       end, [])
 
       assert Agent.get(agent, fn count -> count end) == 1
@@ -86,7 +91,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
         }
       ]
       ) do
-      Fuse.call(@request, fn _request -> {:error, "error"} end, [])
+      Fuse.call(@request, fn _request -> {:error, %RequestError{}} end, [])
       assert called :fuse.melt("json_api_client")
     end
   end
@@ -133,7 +138,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
         }
       ]
       ) do
-      Fuse.call(request, fn _env -> nil end, @options)
+      Fuse.call(request, fn _env -> {:ok, %Response{}} end, @options)
     end
   end
 
@@ -153,7 +158,7 @@ defmodule JsonApiClient.Middleware.FuseTest do
         }
       ]
       ) do
-      Fuse.call(request, fn _env -> nil end, options)
+      Fuse.call(request, fn _env -> {:ok, %Response{}} end, options)
     end
   end
 end