+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use package name as prefix instead of ExApiClient.
 - set Mix.Project.config[:app] as a default value for user_agent_suffix
 - added possibility to configure HTTP Client Backend
+- Added middleware architecture
+- Added `Fuse` and `StatsTracker` middlewares
 - Bug fix when path generation is a post request
 
 ## [1.0.0] - 2017-10-17